Users/requirements.txt
asgiref==3.5.0
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
generics==2.0.0
pytz==2022.1
sqlparse==0.4.2
